Использую long вместо int (месяцы)

+++ src/main/java/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         SalesStatistics service = new SalesStatistics();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         System.out.println(service.sumOfSales(sales));
         System.out.println(service.averageSumOfSales(sales));
